+++ src/addons/character_dna/release_notes.md
@@ -2,6 +2,7 @@
 
 * Added experimental option to turn off batched evaluations
 * Fixed center eye control baking bug [#308](https://github.com/poly-hammer/character-dna-addon/issues/308)
+* Fixed bug with eye aim control when eyes follow head is false [#309](https://github.com/poly-hammer/character-dna-addon/issues/309)
 
 ## Tests Passing On